backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pillow
python-multipart
